+++ tests/Serializers/GzipTest.php
@@ -47,13 +47,13 @@ class GzipTest extends TestCase
     /**
      * @test
      */
-    public function serializeUnserialize() : void
+    public function serializeDeserialize() : void
     {
         $data = $this->serializer->serialize($this->persistable);
 
         $this->assertInstanceOf(Encoding::class, $data);
 
-        $persistable = $this->serializer->unserialize($data);
+        $persistable = $this->serializer->deserialize($data);
 
         $this->assertInstanceOf(DummyClassifier::class, $persistable);
         $this->assertInstanceOf(Persistable::class, $persistable);